requirements.txt
Django==3.2.12
python-dateutil==2.8.2
django-rest-framework==0.1.0
drf-spectacular==0.25.1